requirements.txt
certifi==2025.10.5
cloudinary==1.44.1
python-dotenv==1.1.1
six==1.17.0
urllib3==2.5.0
mysql-connector-python==9.4.0